filled triangles with two sliders for U,V polylines granularity
