+++ main.js
@@ -1,3 +1,5 @@
+import { SurfaceModel } from "./geometry.js";
+
 let gl; // The webgl context.
 let surface; // A surface model
 let program; // A shader program
@@ -5,105 +7,6 @@ let ball; // A SimpleRotator object that lets the user rotate the view by mouse.
 
 function deg2rad(angle) {
 	return (angle * Math.PI) / 180;
-}
-
-class SurfaceModel {
-	constructor(name) {
-		this.name = name;
-		this.uLines = [];
-		this.vLines = [];
-		this.vertexBuffer = null;
-	}
-
-	createSurfaceData() {
-		const uMin = -1;
-		const uMax = 1;
-		const vMin = 0.2;
-		const vMax = 1;
-		const uSteps = 50;
-		const vSteps = 50;
-
-		const du = (uMax - uMin) / uSteps;
-		const dv = (vMax - vMin) / vSteps;
-
-		// Generate U and V lines
-		for (let uIndex = 0; uIndex <= uSteps; uIndex++) {
-			const u = uMin + uIndex * du;
-			const uLine = [];
-			for (let vIndex = 0; vIndex <= vSteps; vIndex++) {
-				const v = vMin + vIndex * dv;
-
-				// Parametric equation for Richmond's Minimal Surface
-				const x =
-					(-3 * u - u ** 5 + 2 * u ** 3 * v ** 2 + 3 * u * v ** 4) /
-					(6 * (u ** 2 + v ** 2));
-				const y =
-					(-3 * v - 3 * u ** 4 * v - 2 * u ** 2 * v ** 3 + v ** 5) /
-					(6 * (u ** 2 + v ** 2));
-				const z = u;
-
-				uLine.push([x, y, z]);
-			}
-			this.uLines.push(uLine);
-		}
-
-		// Generate V lines by transposing the uLines array
-		for (let vIndex = 0; vIndex <= vSteps; vIndex++) {
-			const vLine = [];
-			for (let uIndex = 0; uIndex <= uSteps; uIndex++) {
-				vLine.push(this.uLines[uIndex][vIndex]);
-			}
-			this.vLines.push(vLine);
-		}
-	}
-
-	getVertices() {
-		const vertices = [];
-		// Add U lines
-		for (const uLine of this.uLines) {
-			for (const point of uLine) {
-				vertices.push(...point);
-			}
-		}
-		// Add V lines
-		for (const vLine of this.vLines) {
-			for (const point of vLine) {
-				vertices.push(...point);
-			}
-		}
-		return vertices;
-	}
-
-	initBuffer(gl) {
-		const vertices = this.getVertices();
-		this.vertexBuffer = gl.createBuffer();
-		gl.bindBuffer(gl.ARRAY_BUFFER, this.vertexBuffer);
-		gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(vertices), gl.STATIC_DRAW);
-		this.count = vertices.length / 3;
-	}
-
-	draw(gl, program) {
-		gl.bindBuffer(gl.ARRAY_BUFFER, this.vertexBuffer);
-		gl.vertexAttribPointer(program.vertexAttrib, 3, gl.FLOAT, false, 0, 0);
-		gl.enableVertexAttribArray(program.vertexAttrib);
-
-		const pointsPerLine = this.uLines[0].length;
-		const totalUPoints = this.uLines.length * pointsPerLine;
-
-		// Draw U lines
-		for (let i = 0; i < this.uLines.length; i++) {
-			gl.drawArrays(gl.LINE_STRIP, i * pointsPerLine, pointsPerLine);
-		}
-
-		// Draw V lines
-		for (let i = 0; i < this.vLines.length; i++) {
-			gl.drawArrays(
-				gl.LINE_STRIP,
-				totalUPoints + i * this.uLines.length,
-				this.uLines.length,
-			);
-		}
-	}
 }
 
 class ShaderProgram {
@@ -215,3 +118,23 @@ function init() {
 
 	draw();
 }
+
+// add event listener on DOMContentLoaded event
+document.addEventListener("DOMContentLoaded", init);
+
+// handle u,v steppers values change
+const uStepper = document.getElementById("u-stepper");
+const vStepper = document.getElementById("v-stepper");
+uStepper.addEventListener("change", init);
+vStepper.addEventListener("change", init);
+
+// update u,v steppers counter values on change
+const updateCounter = (e) => {
+	const input = e.target;
+	if (!input) return;
+	const container = input.parentNode;
+	const counter = container.querySelector(".form-input__value");
+	counter.textContent = input.value;
+};
+uStepper.addEventListener("change", updateCounter);
+vStepper.addEventListener("change", updateCounter);
